// src/features/registerDendoOutift/types/types.ts
export type RegisterOutfitValues = {
  title: string;
  keywords: string;
  description: string;
  imageUrl: string | null;
  LIGHTTOPS: string | string[] | boolean | undefined;
  HEAVYTOPS: string | string[] | boolean | undefined;
  OUTERWEAR: string | string[] | boolean | undefined;
  BOTTOMS: string | string[] | boolean | undefined;
  SHOES: string | string[] | boolean | undefined;
  ACCESSORIES: string | string[] | boolean | undefined;
};